requirements.txt
# requirements.txt
torch==2.3.0
torchvision==0.18.0
pytorch-lightning==2.2.1
diffusers==0.27.2
transformers==4.39.3
accelerate==0.28.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
tqdm==4.66.2
einops
xformers==0.0.26.post1
peft==0.10.0
huggingface-hub==0.25.2
